+++ src/DOM/events/linkEvent.ts
@@ -1,3 +1,5 @@
-export default function linkEvent(data: any, event: Function) {
+import { LinkedEventData } from '../../core/structures';
+
+export default function linkEvent(data: any, event: Function): LinkedEventData{
 	return { data, event };
 }
