+++ src/typings/inferno-create-class.d.ts
@@ -1,0 +1,3 @@
+declare module 'inferno-create-class' {
+	export default function createClass(component: any): any;
+}
