+++ src/typings/inferno-create-element.d.ts
@@ -1,0 +1,3 @@
+declare module 'inferno-create-element' {
+	export default function createElement(component: any, props: any, ...children): any;
+}
