+++ src/typings/inferno-hyperscript.d.ts
@@ -1,0 +1,3 @@
+declare module 'inferno-hyperscript' {
+	export default function hyperscript(tag: any, props?: any, ...children): any;
+}
