+++ src/typings/inferno-server.d.ts
@@ -1,0 +1,4 @@
+declare module 'inferno-server' {
+	export function renderToStaticMarkup(...rest);
+	export default function renderToString(...rest);
+}
